Use Travis to auto-deploy Middleman
Why:

We want to automatically deploy any changes to the master branch to the
live site. I've seen it done with Wercker, but we don't need a new tool
in the kit.

We followed this blog post to set up Travis to deploy automatically for
us.

https://jahed.io/blog/2015/05/31/
  automating-deployments-to-github-pages-with-middleman-and-travis-ci/

This change addresses the need by:

* Adding and activating middleman-deploy (untested).

* Setting up a script for Travis to run to deploy changes.